+++ module/Batch/src/Command/ShortlinkStats/ShortlinkStatsFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\ShortlinkStats;
+
+class ShortlinkStatsFactory {
+    public function __invoke($controllers): ShortlinkStats
+    {
+        return new ShortlinkStats($controllers->get('faucetdev'));
+    }
+}