+++ module/Batch/src/Command/UserAccountStats/UserAccountStatsFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Batch\Command\UserAccountStats;
+
+class UserAccountStatsFactory
+{
+    public function __invoke($controllers)
+    {
+        return new UserAccountStats($controllers->get('faucetdev'));
+    }
+}