+++ module/Batch/src/Command/UserDailyStats/UserDailyStatsFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Batch\Command\UserDailyStats;
+
+class UserDailyStatsFactory
+{
+    public function __invoke($controllers)
+    {
+        return new UserDailyStats($controllers->get('faucetdev'));
+    }
+}